fix: split spy relocate logic tests under 400-line gate (#4219)
