+++ packages/colonizethis_logic/test/spy_relocate_intel_test.dart
@@ -2,78 +2,41 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-Game _twoProvinceSpyGame() {
-  return Game(
-    id: 'g1',
-    players: const [
-      Player(id: 'h1', displayName: 'Human', isHuman: true),
-      Player(id: 'gp2', displayName: 'Rival', isHuman: false),
-    ],
-    worldState: WorldState(
-      turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: RegionData(
-        provinces: const [
-          Province(
-            id: 'oldWorld|p1',
-            regionId: 'oldWorld',
-            displayName: 'Home',
-            ownerId: 'h1',
-          ),
-          Province(
-            id: 'oldWorld|p2',
-            regionId: 'oldWorld',
-            displayName: 'Rival Land',
-            ownerId: 'gp2',
-          ),
-        ],
-        units: [
-          Unit(
-            id: 'spy1',
-            type: kUnitTypeSpy,
-            ownerId: 'h1',
-            locationProvinceId: 'oldWorld|p2',
-            tileKey: 'oldWorld|p2|0|0',
-          ),
-        ],
-      ),
-      newWorld: const RegionData(),
-    ),
-  );
-}
+import 'spy_relocate_intel_fixtures.dart';
 
 void main() {
   group('isForeignProvinceForPlayer', () {
     test('returns true for rival-owned province', () {
-      final game = _twoProvinceSpyGame();
+      final game = spyRelocateTwoProvinceGame();
       expect(
         isForeignProvinceForPlayer(
           game: game,
           prefixedProvinceId: 'oldWorld|p2',
-          humanPlayerId: 'h1',
+          humanPlayerId: kSpyRelocateHumanId,
         ),
         isTrue,
       );
     });
 
     test('returns false for human-owned province', () {
-      final game = _twoProvinceSpyGame();
+      final game = spyRelocateTwoProvinceGame();
       expect(
         isForeignProvinceForPlayer(
           game: game,
           prefixedProvinceId: 'oldWorld|p1',
-          humanPlayerId: 'h1',
+          humanPlayerId: kSpyRelocateHumanId,
         ),
         isFalse,
       );
     });
 
     test('returns false for unknown province', () {
-      final game = _twoProvinceSpyGame();
+      final game = spyRelocateTwoProvinceGame();
       expect(
         isForeignProvinceForPlayer(
           game: game,
           prefixedProvinceId: 'oldWorld|missing',
-          humanPlayerId: 'h1',
+          humanPlayerId: kSpyRelocateHumanId,
         ),
         isFalse,
       );
@@ -81,15 +44,13 @@ void main() {
   });
 
   group('countOwnSpiesProjectedInProvince', () {
-    const humanId = 'h1';
-
     test('counts Spies projected in foreign province', () {
-      final game = _twoProvinceSpyGame();
+      final game = spyRelocateTwoProvinceGame();
       expect(
         countOwnSpiesProjectedInProvince(
           game: game,
           orders: const Orders(),
-          humanPlayerId: humanId,
+          humanPlayerId: kSpyRelocateHumanId,
           prefixedProvinceId: 'oldWorld|p2',
         ),
         1,
@@ -97,10 +58,10 @@ void main() {
     });
 
     test('uses pending move destination for projection', () {
-      final game = _twoProvinceSpyGame();
+      final game = spyRelocateTwoProvinceGame();
       const orders = Orders(
         moveOrdersByPlayerId: {
-          humanId: [
+          kSpyRelocateHumanId: [
             MoveOrder(
               unitId: 'spy1',
               destinationTileKey: 'oldWorld|p1|0|0',
@@ -112,7 +73,7 @@ void main() {
         countOwnSpiesProjectedInProvince(
           game: game,
           orders: orders,
-          humanPlayerId: humanId,
+          humanPlayerId: kSpyRelocateHumanId,
           prefixedProvinceId: 'oldWorld|p2',
         ),
         0,
@@ -121,7 +82,7 @@ void main() {
         countOwnSpiesProjectedInProvince(
           game: game,
           orders: orders,
-          humanPlayerId: humanId,
+          humanPlayerId: kSpyRelocateHumanId,
           prefixedProvinceId: 'oldWorld|p1',
         ),
         1,
@@ -130,12 +91,10 @@ void main() {
   });
 
   group('applySpyRelocateMoveToOrders', () {
-    const humanId = 'h1';
-
     test('stages move and clears conflicting work order', () {
       const orders = Orders(
         workOrdersByPlayerId: {
-          humanId: [
+          kSpyRelocateHumanId: [
             WorkOrder(
               unitId: 'spy1',
               target: kWorkTargetCounterSpy,
@@ -146,15 +105,15 @@ void main() {
       );
       final next = applySpyRelocateMoveToOrders(
         orders: orders,
-        humanPlayerId: humanId,
+        humanPlayerId: kSpyRelocateHumanId,
         spyUnitId: 'spy1',
         destinationTileKey: 'oldWorld|p1|1|0',
       );
-      expect(next.workOrdersByPlayerId[humanId], isEmpty);
+      expect(next.workOrdersByPlayerId[kSpyRelocateHumanId], isEmpty);
       expect(
         pendingCivilianMoveForUnit(
           orders: next,
-          humanPlayerId: humanId,
+          humanPlayerId: kSpyRelocateHumanId,
           unitId: 'spy1',
         ),
         const MoveOrder(
@@ -167,7 +126,7 @@ void main() {
     test('replaces prior draft move for same Spy', () {
       const orders = Orders(
         moveOrdersByPlayerId: {
-          humanId: [
+          kSpyRelocateHumanId: [
             MoveOrder(
               unitId: 'spy1',
               destinationTileKey: 'oldWorld|p1|0|0',
@@ -177,25 +136,24 @@ void main() {
       );
       final next = applySpyRelocateMoveToOrders(
         orders: orders,
-        humanPlayerId: humanId,
+        humanPlayerId: kSpyRelocateHumanId,
         spyUnitId: 'spy1',
         destinationTileKey: 'oldWorld|p1|2|0',
       );
-      expect(next.moveOrdersByPlayerId[humanId], hasLength(1));
+      expect(next.moveOrdersByPlayerId[kSpyRelocateHumanId], hasLength(1));
       expect(
-        next.moveOrdersByPlayerId[humanId]!.single.destinationTileKey,
+        next.moveOrdersByPlayerId[kSpyRelocateHumanId]!.single
+            .destinationTileKey,
         'oldWorld|p1|2|0',
       );
     });
   });
 
   group('removePendingCivilianMoveForUnit', () {
-    const humanId = 'h1';
-
     test('removes pending move when present', () {
       const orders = Orders(
         moveOrdersByPlayerId: {
-          humanId: [
+          kSpyRelocateHumanId: [
             MoveOrder(
               unitId: 'spy1',
               destinationTileKey: 'oldWorld|p1|0|0',
@@ -205,10 +163,10 @@ void main() {
       );
       final next = removePendingCivilianMoveForUnit(
         orders: orders,
-        humanPlayerId: humanId,
+        humanPlayerId: kSpyRelocateHumanId,
         unitId: 'spy1',
       );
-      expect(next.moveOrdersByPlayerId[humanId], isEmpty);
+      expect(next.moveOrdersByPlayerId[kSpyRelocateHumanId], isEmpty);
     });
 
     test('returns same orders when no pending move', () {
@@ -217,280 +175,12 @@ void main() {
         identical(
           removePendingCivilianMoveForUnit(
             orders: orders,
-            humanPlayerId: humanId,
+            humanPlayerId: kSpyRelocateHumanId,
             unitId: 'spy1',
           ),
           orders,
         ),
         isTrue,
-      );
-    });
-  });
-
-  group('spyLeaveIntelWarningNeeded', () {
-    const humanId = 'h1';
-
-    test('warns when last Spy would leave foreign province', () {
-      final game = _twoProvinceSpyGame();
-      expect(
-        spyLeaveIntelWarningNeeded(
-          game: game,
-          orders: const Orders(),
-          humanPlayerId: humanId,
-          spyUnitId: 'spy1',
-          newDestinationTileKey: 'oldWorld|p1|0|0',
-        ),
-        isTrue,
-      );
-    });
-
-    test('no warning when another Spy remains in foreign province', () {
-      final game = Game(
-        id: 'g1',
-        players: const [
-          Player(id: 'h1', displayName: 'Human', isHuman: true),
-          Player(id: 'gp2', displayName: 'Rival', isHuman: false),
-        ],
-        worldState: WorldState(
-          turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: const [
-              Province(
-                id: 'oldWorld|p1',
-                regionId: 'oldWorld',
-                displayName: 'Home',
-                ownerId: 'h1',
-              ),
-              Province(
-                id: 'oldWorld|p2',
-                regionId: 'oldWorld',
-                displayName: 'Rival Land',
-                ownerId: 'gp2',
-              ),
-            ],
-            units: [
-              Unit(
-                id: 'spy1',
-                type: kUnitTypeSpy,
-                ownerId: 'h1',
-                locationProvinceId: 'oldWorld|p2',
-                tileKey: 'oldWorld|p2|0|0',
-              ),
-              Unit(
-                id: 'spy2',
-                type: kUnitTypeSpy,
-                ownerId: 'h1',
-                locationProvinceId: 'oldWorld|p2',
-                tileKey: 'oldWorld|p2|1|0',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-      );
-      expect(
-        spyLeaveIntelWarningNeeded(
-          game: game,
-          orders: const Orders(),
-          humanPlayerId: humanId,
-          spyUnitId: 'spy1',
-          newDestinationTileKey: 'oldWorld|p1|0|0',
-        ),
-        isFalse,
-      );
-    });
-
-    test(
-      'warns when other Spy pending move already vacates foreign province',
-      () {
-        final game = Game(
-          id: 'g1',
-          players: const [
-            Player(id: 'h1', displayName: 'Human', isHuman: true),
-            Player(id: 'gp2', displayName: 'Rival', isHuman: false),
-          ],
-          worldState: WorldState(
-            turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-            oldWorld: RegionData(
-              provinces: const [
-                Province(
-                  id: 'oldWorld|p1',
-                  regionId: 'oldWorld',
-                  displayName: 'Home',
-                  ownerId: 'h1',
-                ),
-                Province(
-                  id: 'oldWorld|p2',
-                  regionId: 'oldWorld',
-                  displayName: 'Rival Land',
-                  ownerId: 'gp2',
-                ),
-              ],
-              units: [
-                Unit(
-                  id: 'spy1',
-                  type: kUnitTypeSpy,
-                  ownerId: 'h1',
-                  locationProvinceId: 'oldWorld|p2',
-                  tileKey: 'oldWorld|p2|0|0',
-                ),
-                Unit(
-                  id: 'spy2',
-                  type: kUnitTypeSpy,
-                  ownerId: 'h1',
-                  locationProvinceId: 'oldWorld|p2',
-                  tileKey: 'oldWorld|p2|1|0',
-                ),
-              ],
-            ),
-            newWorld: const RegionData(),
-          ),
-        );
-        const orders = Orders(
-          moveOrdersByPlayerId: {
-            'h1': [
-              MoveOrder(
-                unitId: 'spy2',
-                destinationTileKey: 'oldWorld|p1|0|0',
-              ),
-            ],
-          },
-        );
-        expect(
-          spyLeaveIntelWarningNeeded(
-            game: game,
-            orders: orders,
-            humanPlayerId: humanId,
-            spyUnitId: 'spy1',
-            newDestinationTileKey: 'oldWorld|p1|1|0',
-          ),
-          isTrue,
-        );
-      },
-    );
-
-    test(
-      'no warning when other Spy pending move stays in foreign province',
-      () {
-        final game = Game(
-          id: 'g1',
-          players: const [
-            Player(id: 'h1', displayName: 'Human', isHuman: true),
-            Player(id: 'gp2', displayName: 'Rival', isHuman: false),
-          ],
-          worldState: WorldState(
-            turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-            oldWorld: RegionData(
-              provinces: const [
-                Province(
-                  id: 'oldWorld|p1',
-                  regionId: 'oldWorld',
-                  displayName: 'Home',
-                  ownerId: 'h1',
-                ),
-                Province(
-                  id: 'oldWorld|p2',
-                  regionId: 'oldWorld',
-                  displayName: 'Rival Land',
-                  ownerId: 'gp2',
-                ),
-              ],
-              units: [
-                Unit(
-                  id: 'spy1',
-                  type: kUnitTypeSpy,
-                  ownerId: 'h1',
-                  locationProvinceId: 'oldWorld|p2',
-                  tileKey: 'oldWorld|p2|0|0',
-                ),
-                Unit(
-                  id: 'spy2',
-                  type: kUnitTypeSpy,
-                  ownerId: 'h1',
-                  locationProvinceId: 'oldWorld|p2',
-                  tileKey: 'oldWorld|p2|1|0',
-                ),
-              ],
-            ),
-            newWorld: const RegionData(),
-          ),
-        );
-        const orders = Orders(
-          moveOrdersByPlayerId: {
-            'h1': [
-              MoveOrder(
-                unitId: 'spy2',
-                destinationTileKey: 'oldWorld|p2|2|0',
-              ),
-            ],
-          },
-        );
-        expect(
-          spyLeaveIntelWarningNeeded(
-            game: game,
-            orders: orders,
-            humanPlayerId: humanId,
-            spyUnitId: 'spy1',
-            newDestinationTileKey: 'oldWorld|p1|0|0',
-          ),
-          isFalse,
-        );
-      },
-    );
-
-    test('returns false for unknown unit', () {
-      final game = _twoProvinceSpyGame();
-      expect(
-        spyLeaveIntelWarningNeeded(
-          game: game,
-          orders: const Orders(),
-          humanPlayerId: humanId,
-          spyUnitId: 'missing',
-          newDestinationTileKey: 'oldWorld|p1|0|0',
-        ),
-        isFalse,
-      );
-    });
-
-    test('returns false when relocating from owned province', () {
-      final game = Game(
-        id: 'g1',
-        players: const [
-          Player(id: 'h1', displayName: 'Human', isHuman: true),
-        ],
-        worldState: WorldState(
-          turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: const [
-              Province(
-                id: 'oldWorld|p1',
-                regionId: 'oldWorld',
-                displayName: 'Home',
-                ownerId: 'h1',
-              ),
-            ],
-            units: [
-              Unit(
-                id: 'spy1',
-                type: kUnitTypeSpy,
-                ownerId: 'h1',
-                locationProvinceId: 'oldWorld|p1',
-                tileKey: 'oldWorld|p1|0|0',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-      );
-      expect(
-        spyLeaveIntelWarningNeeded(
-          game: game,
-          orders: const Orders(),
-          humanPlayerId: humanId,
-          spyUnitId: 'spy1',
-          newDestinationTileKey: 'oldWorld|p1|1|0',
-        ),
-        isFalse,
       );
     });
   });
